+++ packages/mds-compliance-engine/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@mds-core/mds-compliance-batch-processor",
+  "name": "@mds-core/mds-compliance-engine",
   "version": "0.1.27",
   "description": "MDS Compliance Tool",
   "main": "dist/index.js",
@@ -9,7 +9,6 @@
   ],
   "scripts": {
     "build": "tsc --build tsconfig.build.json",
-    "start": "PATH_PREFIX=/compliance yarn watch server",
     "test": "yarn test:eslint && yarn test:unit",
     "test:eslint": "eslint --ignore-path ../../.gitignore '**/*.ts'",
     "test:unit": "PATH_PREFIX=/compliance DOTENV_CONFIG_PATH=../../.env nyc ts-mocha --project ../../tsconfig.json --exit",
@@ -18,7 +17,7 @@
   },
   "keywords": [
     "mds",
-    "compliance"
+    "compliance-engine"
   ],
   "author": "City of Los Angeles.",
   "license": "Apache-2.0",
@@ -29,10 +28,11 @@
     "@mds-core/mds-providers": "0.1.26",
     "@mds-core/mds-schema-validators": "0.1.2",
     "@mds-core/mds-stream": "0.1.26",
+    "@mds-core/mds-test-data": "0.1.26",
+    "@types/moment-timezone": "0.5.13",
     "@mds-core/mds-types": "0.1.23",
     "@mds-core/mds-utils": "0.1.26",
     "@mds-core/mds-compliance-service": "0.1.0",
-    "@types/moment-timezone": "0.5.13",
     "@types/yargs": "15.0.5",
     "fs": "0.0.1-security",
     "moment-timezone": "0.5.31",
